[ISSUE #9627]♻️Partition RemotingCommand state and custom-header implementations
